Add files via upload
PHPMailer-master library file has been cited inside of contactUs.php. this contains NO html as this is the functionality which sends an email from the company email to the customer email. which also saves the users credentials in a database along with their message. NOTE: DO NOT CHANGE THE VARIABLE NAMES