Add files via upload
book appointment function, it allows the user to choose a date and time and enter their credentials along with a submit button. their credentials, time and date they chose is stored into a database called appointments.sql.
Then it sends an email to the user along with a unique 8 digit reference number.